more cypher all the time

+++ net.locosoft.fold.channel.fold/src/net/locosoft/fold/channel/fold/IFoldChannel.java
@@ -15,6 +15,6 @@ import net.locosoft.fold.channel.IChannel;
 
 public interface IFoldChannel extends IChannel {
 
-	int getStartCount();
+	long getStartCount();
 
 }
